requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
Werkzeug
PyJWT